satellite/admin/ui: Migrate to SvelteKit
Migrate the satellite admin UI web app from the Svelte templated used to
generate a Svelte App scaffolding to SvelteKit.

The application has not changed that much because we don't have more
than one route, however the migration allows to use the new tooling
rather than Rollup.

On the other hand, it's better to use SvelteKit in case that in the
future the web app grows and we start to create routes, etc.

Change-Id: Ifa6736c13585708337f6c5a59388077b784eaddd